feat: use separated style file in content and options
- disable preflight in content css
