+++ public/options.html
@@ -1,14 +1,11 @@
-<!DOCTYPE html>
+<!doctype html>
 <html>
   <head>
     <meta charset="utf-8" />
     <link rel="stylesheet" href="dist/options.css" />
     <title>Summer Web Clipper: Options</title>
   </head>
-  <body
-    class="swc:container swc:preflight swc:mx-auto swc:px-20"
-    style="font-size: 100%"
-  >
+  <body class="swc:container swc:mx-auto swc:px-20">
     <div id="root"></div>
     <script src="./dist/options.js"></script>
   </body>
